Padronizar formato da moeda

+++ laravel/database/migrations/2023_11_28_201224_create_contas_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('contas', function (Blueprint $table) {
             $table->id('conta_id');
-            $table->float('saldo');
+            $table->decimal('saldo', 12, 2);
         });
     }
